+++ src/main/java/kraheja/payroll/bean/response/EmpeducationResponseBean.java
@@ -38,7 +38,7 @@ public class EmpeducationResponseBean {
 	private String mainsubjetcs ;
 	private LocalDateTime modifiedon ;
 	private String module ;
-	private Integer percofmarks ;
+	private Double percofmarks ;
 	private String site ;
 	private String userid ;
 }
